feat : 질문(게시글) 삭제시 자신의 댓글만이 존재하는 경우에만 삭제하는 제약 구현 - QnaService#deleteQuestion에서 질문(게시글) 삭제시 다른 사람의 댓글이 존재하는 경우 삭제하지 않도록 검사합니다. - 질문(게시글) 생성시 user의 pk를 저장하지 않았던 버그를 수정하였습니다.

+++ src/main/java/woowa/cafe/service/QnaService.java
@@ -5,6 +5,7 @@ import woowa.cafe.dto.QuestionInfo;
 import woowa.cafe.dto.UpdateQuestionRequest;
 import woowa.cafe.dto.request.CreateQuestionRequest;
 import woowa.cafe.repository.QuestionRepository;
+import woowa.cafe.repository.ReplyRepository;
 import woowa.frame.core.annotation.Component;
 
 import java.util.List;
@@ -13,9 +14,11 @@ import java.util.List;
 public class QnaService {
 
     private final QuestionRepository questionRepository;
+    private final ReplyRepository replyRepository;
 
-    public QnaService(QuestionRepository questionRepository) {
+    public QnaService(QuestionRepository questionRepository, ReplyRepository replyRepository) {
         this.questionRepository = questionRepository;
+        this.replyRepository = replyRepository;
     }
 
     public void createQna(CreateQuestionRequest request) {
@@ -72,7 +75,12 @@ public class QnaService {
         if (question == null) return false;
         if (!question.getUserId().equals(userId)) return false;
 
-        questionRepository.deleteById(questionId);
-        return true;
+        if (replyRepository.existsByQuestionIdAndNotUserId(questionId, userId)) {
+            return false;
+        } else {
+            questionRepository.deleteById(questionId);
+            replyRepository.deleteByQuestionId(questionId);
+            return true;
+        }
     }
 }
